Database moved to root directory
The prior default location of the user's "My Documents" folder is not reliable as the OneDrive backup feature when enabled moves all files from My Documents to an equivalent under OneDrive. This leads to the application creating a new sut.db database file in My Documents which is then moved arbitrarily by OneDrive causing fragmentation of work units across multiple database files. This change stores the database in the root directory where the binaries are stored.